Clean up PatientHome - NoteWriter needs to be fixed
