+++ src/routes/Patient/Patient.jsx
@@ -3,7 +3,6 @@ import { AppBar, Box, Tab, Tabs, Divider, Drawer, Stack, IconButton} from '@mui/
 import { TabContext, TabList, TabPanel } from '@mui/lab';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import { useTheme } from '@mui/material/styles';
-import {EditorState} from 'draft-js'; // Modifier, ContentState (Maybe)
 import { Icon } from '../../components/ui/Core.jsx';
 
 import { usePatientMRN, useEncounterID } from '../../util/urlHelpers.js';
@@ -21,371 +20,19 @@ import Medications from './routes/Medications/Medications.jsx';
 import ResultsReview from "./routes/Results/Results.jsx";
 import Pdmp from './routes/PDMP/PDMP.jsx';
 import Immunizations from './routes/Immunizations/Immunizations.jsx';
-import Allergies from './routes/Allergies/Allergies.jsx';
+import { Allergies } from './routes/Allergies/Allergies.jsx';
 import Chat from "./routes/Chat/Chat.jsx";
-
-const bodySystems = [
-  {
-    title: 'Constitutional',
-    symptoms: ['Fever', 'Chills', 'Weight loss', 'Malaise/Fatigue', 'Diaphoresis'],
-  },
-  {
-    title: 'Skin',
-    symptoms: ['Rash', 'Itching'],
-  },
-  {
-    title: 'HENT',
-    symptoms: [
-      'Hearing loss',
-      'Tinnitus',
-      'Ear pain',
-      'Ear discharge',
-      'Nosebleeds',
-      'Congestion',
-      'Sinus pain',
-      'Stridor',
-      'Sore throat',
-    ],
-  },
-  {
-    title: 'Eyes',
-    symptoms: [
-      'Blurred vision',
-      'Double vision',
-      'Photophobia',
-      'Eye pain',
-      'Eye discharge',
-      'Eye redness',
-    ],
-  },
-  {
-    title: 'Cardiovascular',
-    symptoms: [
-      'Chest pain',
-      'Palpitations',
-      'Orthopnea',
-      'Claudication',
-      'Leg swelling',
-      'PND',
-    ],
-  },
-  {
-    title: 'Respiratory',
-    symptoms: [
-      'Cough',
-      'Hemoptysis',
-      'Sputum production',
-      'Shortness of breath',
-      'Wheezing',
-    ],
-  },
-  {
-    title: 'GI',
-    symptoms: [
-      'Heartburn',
-      'Nausea',
-      'Vomiting',
-      'Abdominal pain',
-      'Diarrhea',
-      'Constipation',
-      'Blood in stool',
-      'Melena',
-    ],
-  },
-  {
-    title: 'GU',
-    symptoms: [
-      'Dysuria',
-      'Urgency',
-      'Frequency',
-      'Hematuria',
-      'Flank pain',
-    ],
-  },
-  {
-    title: 'Musculoskeletal',
-    symptoms: [
-      'Myalgias',
-      'Neck pain',
-      'Back pain',
-      'Joint pain',
-      'Falls',
-    ],
-  },
-  {
-    title: 'Endo/Heme/Allergy',
-    symptoms: [
-      'Easy bruising/bleeding',
-      'Environmental allergies',
-      'Polydipsia',
-    ],
-  },
-  {
-    title: 'Neurological',
-    symptoms: [
-      'Dizziness',
-      'Headaches',
-      'Tingling',
-      'Tremor',
-      'Sensory change',
-      'Speech change',
-      'Focal weakness',
-      'Weakness',
-      'Seizures',
-      'LOC',
-    ],
-  },
-  {
-    title: 'Psychiatric',
-    symptoms: [
-      'Depression',
-      'Suicidal ideas',
-      'Substance abuse',
-      'Hallucinations',
-      'Nervous/Anxious',
-      'Insomnia',
-      'Memory loss',
-    ],
-  },
-];
-
-const physicalExamBodySystems = [
-  {
-    title: 'Constitutional',
-    subsections: [
-      {
-        // subsectionTitle: 'General', 
-        checkboxes: ['Alert', 'Normal Weight', 'Normal Appearance', 'Obese'], 
-        symptoms: ['Acute Distress', 'Ill-Appearing', 'Toxic Appearing', 'Diaphoretic'], 
-      }, 
-    ],
-  },
-  { 
-    title: 'Neck', 
-    subsections: [ 
-      { 
-        checkboxes: ['ROM Normal', 'Supple'], 
-        symptoms: ['Neck Rigidity', 'Tenderness', 'Cervical Adenopathy', 'Carotid Bruit'], 
-      }, 
-    ],
-  },
-  {
-    title: 'Skin',
-    subsections: [
-      {
-        checkboxes: ['Warm', 'Dry', 'Normal Color'],
-        symptoms: ['Rash', 'Erythema', 'Pallor', 'Cyanosis'],
-      },
-    ],
-  },
-  {
-    title: 'Respiratory',
-    subsections: [
-      {
-        subsectionTitle: 'Breath Sounds',
-        checkboxes: ['Clear to Auscultation Bilaterally'],
-        symptoms: ['Wheezes', 'Rales', 'Rhonchi'],
-      },
-      {
-        subsectionTitle: 'Respiratory Effort',
-        checkboxes: ['Normal Effort', 'No Accessory Muscle Use'],
-        symptoms: ['Labored Breathing', 'Retractions'],
-      },
-    ],
-  },
-  {
-    title: 'Genitourinary / Anorectal',
-    subsections: [
-      {
-        subsectionTitle: 'General',
-        checkboxes: ['Normal External Appearance'],
-        symptoms: ['Lesions', 'Discharge'],
-      },
-      {
-        subsectionTitle: 'Anorectal',
-        checkboxes: ['Normal Tone'],
-        symptoms: ['Hemorrhoids', 'Fissures'],
-      },
-    ],
-  },
-  {
-    title: 'Musculoskeletal',
-    subsections: [
-      {
-        subsectionTitle: 'General',
-        checkboxes: ['Full Range of Motion', 'No Deformity'],
-        symptoms: ['Tenderness', 'Swelling', 'Limited ROM'],
-      },
-      {
-        subsectionTitle: 'Spine',
-        checkboxes: ['Normal Curvature'],
-        symptoms: ['Tenderness', 'Kyphosis', 'Scoliosis'],
-      },
-    ],
-  },
-  {
-    title: 'Neurological',
-    subsections: [
-      {
-        subsectionTitle: 'Mental Status',
-        checkboxes: ['Alert', 'Oriented x3'],
-        symptoms: ['Confusion', 'Lethargy'],
-      },
-      {
-        subsectionTitle: 'Motor Function',
-        checkboxes: ['Normal Strength'],
-        symptoms: ['Weakness', 'Tremor', 'Involuntary Movements'],
-      },
-      {
-        subsectionTitle: 'Reflexes',
-        checkboxes: ['Normal Reflexes'],
-        symptoms: ['Hyperreflexia', 'Hyporeflexia'],
-      },
-    ],
-  },
-  {
-    title: 'Psychiatric',
-    subsections: [
-      {
-        subsectionTitle: 'Affect',
-        checkboxes: ['Normal Affect'],
-        symptoms: ['Flat', 'Anxious', 'Depressed'],
-      },
-      {
-        subsectionTitle: 'Behavior',
-        checkboxes: ['Cooperative'],
-        symptoms: ['Agitated', 'Withdrawn'],
-      },
-    ],
-  },
-  {
-    title: 'Gastrointestinal',
-    subsections: [
-      {
-        subsectionTitle: 'Inspection',
-        checkboxes: ['Non-Distended'],
-        symptoms: ['Scars', 'Striae'],
-      },
-      {
-        subsectionTitle: 'Palpation',
-        checkboxes: ['Soft', 'Non-Tender'],
-        symptoms: ['Tenderness', 'Guarding', 'Rebound'],
-      },
-      {
-        subsectionTitle: 'Bowel Sounds',
-        checkboxes: ['Normal Bowel Sounds'],
-        symptoms: ['Hyperactive', 'Hypoactive', 'Absent'],
-      },
-    ],
-  },
-  {
-    title: 'HEENT',
-    subsections: [
-      {
-        subsectionTitle: 'Head',
-        checkboxes: ['Normocephalic', 'Atraumatic'],
-        symptoms: [],
-      },
-      {
-        subsectionTitle: 'Eyes',
-        checkboxes: ['PERRL', 'EOM Intact'],
-        symptoms: ['Conjunctivae Normal', 'Scleral Icterus'],
-      },
-      {
-        subsectionTitle: 'Ears (Right)',
-        checkboxes: ['TM Normal', 'Canal Normal', 'External Ear Normal'],
-        symptoms: ['Impacted Cerumen'],
-      },
-      {
-        subsectionTitle: 'Ears (Left)',
-        checkboxes: ['TM Normal', 'Canal Normal', 'External Ear Normal'], 
-        symptoms: ['Impacted Cerumen'], 
-      }, 
-      { 
-        subsectionTitle: 'Nose', 
-        checkboxes: ['Nose Normal'], 
-        symptoms: ['Congestion', 'Rhinorrhea'], 
-      }, 
-      { 
-        subsectionTitle: 'Mouth/Throat', 
-        checkboxes: ['Moist'], 
-        symptoms: ['Clear', 'Exudate', 'Erythema'], 
-      }, 
-    ], 
-  },
-  {
-    title: 'Cardiovascular',
-    subsections: [
-      {
-        subsectionTitle: 'Rate',
-        checkboxes: ['Normal Rate', 'Tachycardia', 'Bradycardia'],
-        symptoms: []
-      },
-      {
-        subsectionTitle: 'Rhythm',
-        checkboxes: ['Regular Rhythm', 'Irregular Rhythm'],
-        symptoms: []
-      },
-      {
-        subsectionTitle: 'Pulses and Heart Sounds',
-        checkboxes: ['Pulses Normal', 'Heart Sounds Normal'],
-        symptoms: ['Murmur', 'Friction Rub', 'Gallop']
-      },
-    ],
-  },
-
-];
-
 
 export const PatientHome = ({ ...props }) => {
   const [enc] = useEncounterID();
+  const [patientMRN] = usePatientMRN();
+  const [patientData, setPatientData] = useState(TEST_PATIENT_INFO({ patientMRN }));
 
   const drawerWidth = 250
   const [tab, setTab] = useState("1")
   const [storyboardOpen, setStoryboardOpen] = useState(true)
+  const isMobile = useMediaQuery(useTheme().breakpoints.down('sm'));
 
-  const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
-
-  const [editorState, setEditorState] = useState(() => EditorState.createEmpty());
-
-  const generateInitialRosState = (systems) => {
-    const initialState = {};
-  
-    systems.forEach(system => {
-      const systemState = {};
-      system.symptoms.forEach(symptom => {
-        systemState[symptom] = null; // Initial state for each symptom is `null`
-      });
-      systemState['custom'] = null; // Add a custom entry for custom ROS text (DescriptionIcon)
-      initialState[system.title.toLowerCase()] = systemState;
-    });
-  
-    return initialState;
-  };
-
-  const generateInitialPEState = (systems) => {
-    // We need a separate function to generate the initial state for the physical exam because it has subsections
-    const initialState = {};
-  
-    systems.forEach(system => {
-      const systemState = {};
-      system.subsections.forEach((subsection, idx) => {
-        const subsectionTitle = subsection.subsectionTitle ? subsection.subsectionTitle.toLowerCase() : '';
-        systemState[subsectionTitle] = {symptoms: null, checkboxes: null}; // Initial state for each subsection is null
-      });
-      initialState[system.title.toLowerCase()] = systemState;
-      systemState['custom'] = null; // Add a custom entry for custom PE text (DescriptionIcon)
-    });
-  
-    return initialState;
-  };
-  
-  
-  const [peState, setPEState] = useState(generateInitialPEState(physicalExamBodySystems));
-  const [rosState, setRosState] = useState(generateInitialRosState(bodySystems));
-  const [patientMRN] = usePatientMRN();
-  const [patientData, setPatientData] = useState(TEST_PATIENT_INFO({ patientMRN }));
   return (
     <Box display="flex" direction="row" sx={{ overflowY: 'hidden', ...props.sx }} {...props}>
       <Drawer
@@ -454,28 +101,11 @@ export const PatientHome = ({ ...props }) => {
             <TabPanel sx={{ p: 0 }} value="5"><Medications /></TabPanel>
             <TabPanel sx={{ p: 0 }} value="6"><Orders /></TabPanel>
             <TabPanel sx={{ p: 0 }} value="7"><OrdersMgmt/></TabPanel>
-            <TabPanel sx={{ p: 0 }} value="8">
-              <NotesTabContent 
-                editorState={editorState}
-                setEditorState={setEditorState}
-                rosState={rosState} 
-                setRosState={setRosState}
-                peState={peState} 
-                setPEState={setPEState}
-                bodySystems={bodySystems}
-                bodySystemsPE={physicalExamBodySystems}
-                // Want to have ROS & Editor State Here as want it preserved as user is going thru the chart to fill in details (do
-                // not want it to reset each time user switches tabs. This may change down the line, but for now, this is the plan)
-              />
-            </TabPanel>
+            <TabPanel sx={{ p: 0 }} value="8"><NotesTabContent /></TabPanel>
             <TabPanel sx={{ p: 0 }} value="9"><ResultsReview/></TabPanel>
             <TabPanel sx={{ p: 0 }} value="10"><Pdmp/></TabPanel>
             <TabPanel sx={{ p: 0 }} value="11"><Immunizations/></TabPanel>
-            <TabPanel sx={{ p: 0 }} value="12"><Allergies
-                patientData={patientData}
-                setPatientData={setPatientData}
-                encounterId={enc}
-              /></TabPanel>
+            <TabPanel sx={{ p: 0 }} value="12"><Allergies /></TabPanel>
             <TabPanel sx={{ p: 0 }} value="13"><Chat /></TabPanel>
           </Box>
         </TabContext>
